test(pvtime): split pvtime tests into functional and performance
The test contends one physical core to increase the steal time, but if
many tests are run in parallel, they all start contending the same core.
Move the tests to the performance folder so that they are run serially.
